requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
automat==25.4.16
billiard==4.2.1
celery==5.5.3
cffi==1.17.1
channels==4.2.2
channels-redis==4.2.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cryptography==45.0.5
daphne==4.2.1
django==5.2.4
django-filter==25.1
django-redis==6.0.0
djangochannelsrestframework==1.3.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
drf-spectacular==0.28.0
hiredis==3.2.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
msgpack==1.1.1
packaging==25.0
pillow==11.3.0
pip==25.1.1
prompt-toolkit==3.0.51
psycopg==3.2.9
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pyjwt==2.9.0
pyopenssl==25.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pyyaml==6.0.2
redis==6.2.0
referencing==0.36.2
rpds-py==0.26.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
twisted==25.5.0
txaio==25.6.1
typing-extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
uv==0.7.17
vine==5.1.0
wcwidth==0.2.13
zope-interface==7.2
